+ Fixes to suppress warnings with gcc compiler + Fixed const pointer to const value syntax + Fixed signed/unsigned comparison + Fixed undefined behavior due to base class pointer calling only parent's destructor (added a virtual destructor) + Removed many unused variables + Removed typedefs in favor of auto from C++ 11

+++ branson_dd/src/load_balance.h
@@ -25,10 +25,10 @@
 #include "photon.h"
 #include "work_packet.h"
 
-//! Load balance the work on all ranks given an array of work packets and 
+//! Load balance the work on all ranks given an array of work packets and
 // census particles
-void load_balance(const int& rank, const int& n_rank, 
-  const uint64_t n_particle_on_rank, std::vector<Work_Packet>& work, 
+void load_balance(const int& rank, const int& n_rank,
+  const uint64_t n_particle_on_rank, std::vector<Work_Packet>& work,
   std::vector<Photon>& census_list, MPI_Types *mpi_types)
 {
   using std::unordered_map;
@@ -37,7 +37,7 @@ void load_balance(const int& rank, const int& n_rank,
   using Constants::photon_tag;
   using Constants::work_tag;
 
-  // get MPI datatypes 
+  // get MPI datatypes
   MPI_Datatype MPI_Particle = mpi_types->get_particle_type();
 
   MPI_Datatype MPI_WPacket = mpi_types->get_work_packet_type();
@@ -50,24 +50,24 @@ void load_balance(const int& rank, const int& n_rank,
   vector<int64_t> domain_delta_p(n_rank, 0);
 
   domain_particles[rank] = n_particle_on_rank;
-  
+
   MPI_Allreduce(MPI_IN_PLACE, &domain_particles[0], n_rank, MPI_UNSIGNED_LONG,
     MPI_SUM, MPI_COMM_WORLD);
 
-  int64_t n_global_particles = 
+  int64_t n_global_particles =
     std::accumulate(domain_particles.begin(), domain_particles.end(),0);
-  
+
   int64_t n_balance_particles = n_global_particles/n_rank;
 
-  for (uint32_t ir=0; ir<n_rank; ir++) 
+  for (uint32_t ir=0; ir<uint32_t(n_rank); ir++)
     domain_delta_p[ir] = domain_particles[ir] - n_balance_particles;
 
   //--------------------------------------------------------------------------//
-  // Determine which ranks to send or reiceve work from  
+  // Determine which ranks to send or reiceve work from
   //--------------------------------------------------------------------------//
 
   // first, ignore particle imbalance on the order of 5%
-  for (uint32_t ir=0; ir<n_rank; ir++) {
+  for (uint32_t ir=0; ir<uint32_t(n_rank); ir++) {
     // ignore particle imbalance on the order of 5%
     if ( domain_delta_p[ir]<0 && domain_particles[ir] > 0.95*n_balance_particles)
       domain_delta_p[ir] = 0;
@@ -76,13 +76,13 @@ void load_balance(const int& rank, const int& n_rank,
   unordered_map<int32_t, int32_t> n_send_rank;
   vector<uint32_t> work_donor_ranks;
 
-  uint64_t max_particles = 0;
+  int64_t max_particles = 0;
   uint32_t max_rank = 0;
 
   double overload_factor = 0.0; // 40%
 
   int send_to_nbr, rank_delta_p, left_node, right_node;
-  for (uint32_t ir=0; ir<n_rank; ir++) {
+  for (uint32_t ir=0; ir<uint32_t(n_rank); ir++) {
     if (domain_particles[ir] > max_particles) {
       max_particles=domain_particles[ir];
       max_rank = ir;
@@ -92,7 +92,7 @@ void load_balance(const int& rank, const int& n_rank,
     // allow overload on donor ranks by effectively reducing the number of
     // particles they can donate (signed integer math)
     if (rank_delta_p>0) {
-      rank_delta_p = domain_particles[ir] - 
+      rank_delta_p = domain_particles[ir] -
         (1.0 + overload_factor)*n_balance_particles;
     }
 
@@ -105,7 +105,7 @@ void load_balance(const int& rank, const int& n_rank,
           if (send_to_nbr > rank_delta_p) send_to_nbr = rank_delta_p;
 
           // if this is your rank, remember it in your map
-          if (rank==ir) n_send_rank[left_node] = send_to_nbr;
+          if (uint32_t(rank)==ir) n_send_rank[left_node] = send_to_nbr;
 
           // if you are going to receive work, remember donor rank
           if (rank==left_node) work_donor_ranks.push_back(ir);
@@ -121,14 +121,14 @@ void load_balance(const int& rank, const int& n_rank,
           if (send_to_nbr > rank_delta_p) send_to_nbr = rank_delta_p;
 
           // if this is your rank, remember it in your map
-          if (rank==ir) n_send_rank[right_node] = send_to_nbr;
+          if (uint32_t(rank)==ir) n_send_rank[right_node] = send_to_nbr;
           // if you are going to receive work, remember donor rank
 
           if (rank==right_node) work_donor_ranks.push_back(ir);
 
           domain_delta_p[right_node]+=send_to_nbr;
           rank_delta_p -= send_to_nbr;
-        }        
+        }
         right_node++;
       } // if ir+dist < n_rank
     } //end while
@@ -165,14 +165,14 @@ void load_balance(const int& rank, const int& n_rank,
       work_buffer[i].resize(n_deficient);
 
       // post work packet receives
-      MPI_Irecv(work_buffer[i].get_buffer(), n_deficient, MPI_WPacket, 
+      MPI_Irecv(work_buffer[i].get_buffer(), n_deficient, MPI_WPacket,
         work_donor_ranks[i], work_tag, MPI_COMM_WORLD, &work_recv_request[i]);
 
       // make buffer maximum possible size
       photon_buffer[i].resize(n_deficient);
 
       // post particle receives
-      MPI_Irecv(photon_buffer[i].get_buffer(), n_deficient, MPI_Particle, 
+      MPI_Irecv(photon_buffer[i].get_buffer(), n_deficient, MPI_Particle,
         work_donor_ranks[i], photon_tag, MPI_COMM_WORLD, &phtn_recv_request[i]);
     }
 
@@ -187,13 +187,13 @@ void load_balance(const int& rank, const int& n_rank,
 
       // add received work to your work
       vector<Work_Packet> temp_work = work_buffer[i].get_object();
-      work.insert(work.begin(), temp_work.begin(), 
+      work.insert(work.begin(), temp_work.begin(),
         temp_work.begin() + n_work_recv[i]);
 
       // add received census photons
       vector<Photon> temp_photons = photon_buffer[i].get_object();
-      census_list.insert(census_list.begin(), temp_photons.begin(), 
-        temp_photons.begin() + n_phtn_recv[i]);  
+      census_list.insert(census_list.begin(), temp_photons.begin(),
+        temp_photons.begin() + n_phtn_recv[i]);
     }
     delete[] work_recv_request;
     delete[] phtn_recv_request;
@@ -221,7 +221,7 @@ void load_balance(const int& rank, const int& n_rank,
 
     // iterate over unordered map and prepare work for other ranks
     for (unordered_map<int32_t,int32_t>::iterator map_itr=n_send_rank.begin();
-      map_itr !=n_send_rank.end(); map_itr++) 
+      map_itr !=n_send_rank.end(); map_itr++)
     {
 
       dest_rank = map_itr->first;
@@ -237,7 +237,7 @@ void load_balance(const int& rank, const int& n_rank,
           leftover_packet = temp_packet.split(temp_n_send);
           work[work.size()-1] = leftover_packet;
         }
-        // otherwise, pop the temp work packet off the stack 
+        // otherwise, pop the temp work packet off the stack
         else work.pop_back();
 
         // add packet to send list
@@ -267,20 +267,20 @@ void load_balance(const int& rank, const int& n_rank,
       MPI_Wait(&phtn_send_request[ireq], MPI_STATUS_IGNORE);
 
       // increment the request index counter
-      ireq++; 
+      ireq++;
     }
 
     // remove census photons that were sent off
-    census_list.erase(census_list.begin() + n_census_remain, 
+    census_list.erase(census_list.begin() + n_census_remain,
       census_list.end());
 
-    // clean up 
+    // clean up
     delete[] work_send_request;
     delete[] phtn_send_request;
-    
+
   } // end if(n_acceptors)
 
-}  
+}
 
 #endif // load_balance_h_
